move material table to its own component
