+++ frontend/src/components/material/MaterialListItemContentNode.vue
