+++ frontend/src/components/material/MaterialListItemPeriod.vue
